pyproject.toml
[project]
name = "dns-sequence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "geoparse>=2.0.4",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotext>=5.3.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "scikit-learn>=1.6.1",
]
